Issue #3: Resource graphs not shown with CentOS 7
The default recipe uses a local home_dir variable which is set to distribution specific value. The local variable is
not passed to template files and for this reason, the RRD directories doesn't fit. The OpenNMS home variable has to
be set in the Vagrant file to /opt/opennms for CentOS or /usr/share/opennms for Debian/Ubuntu Linux distribution.